Added new multimesh double diffusive convection demo drivers, which required fixing an oversight in set_sources.
Tidied some of the code in the convection problems. In the multimesh
case, far fewer functions needed to be overloaded, so the code is now
much shorter and cleaner. 

Updated the appropriate documentation for multimesh boussinesq convection.

Added the required time-dependent version of get_interpolated_values() in
RefineableAdvectionDiffusion<DIM> elements.



git-svn-id: svn+ssh://vummath.ma.man.ac.uk/home/svn-oomphlib/public/trunk@51 595e1d82-adf1-4282-bac7-b6073fa2ca8e