1) Fixed a stupid bug in axi_static_cap.cc demo in which the swirl velocity was never fixed. This lead to a (near) singular matrix that was not always handled gracefully by the linear solver. Note that the error disappeared under valgrind beacause it uses its own implementation of IEEE floating point. In the process of finding the bug, a number of memory leaks in the driver were fixed.
2) Fixed an oversight in VolumeConstraintBoundingElements in the case when 
the pressure that is traded for the volume constraint is an internal 
fluid pressure and the fluid elements interpolate the pressure at nodes. It is
then possible for the traded pressure to be nodal data of the 
VolumeConstraintBoundingElement, in which case adding it as external data leads
to repeated local equation numbers. This causes a fail when PARANOID error
checking is enabled. The case is now handled properly by testing whether
the traded data is nodal data when it is added, which is a bit more expensive, 
but only happens during setup and can be avoided via a boolean argument 
VolumeConstraintBoundingElement::set_volume_constraint_element(). The default
is to do the checking.

3) Removed buoyant_navier_stokes_elements.h from 
multi_physics/marangoni_convection and multi_physics/insoluble_surfactant 
demo drivers. These now link to the elements defined in src/multi_physics.

4) The use of finite differences for off-diagonal blocks of multi physics 
elements was not as efficient as it should have been. If an element is 
defined via

class MultiPhysics : public Physics1, public Physics2 {};

 then a call to Physics1::get_residuals(...) will be identical to a 
call to MultiPhysics::get_residuals(..) because this function is not 
overloaded and always returns all residuals associated with the element. 
Within get_residuals() the function fill_in_contribution_to_residuals(...) 
is called and it is this that it overloaded. Hence in order to compute 
only the residuals associated with Physics1 we must use
Physics1::fill_in_contribution_to_residuals(...) after initialising the 
residuals to zero. All examples of this pattern in the library have now 
been changed.

5) Another related problem in PseudoSolidNodeUpdateElements was that the
finite-difference loop to determine the shape derivatives was actually 
replacing the entries in all rows, overwriting the diagonal contribution 
from the underlying solid elements. This has now been optimised by 
pre-calculating all solid degrees of freedom and only adding the result of the
finite difference calculations to dofs that are not solid dofs. In order to
handle the case of solid pressures certain functions have been moved into the
PVDEquations base class so that they can be used for all SolidElements. 



git-svn-id: svn+ssh://vummath.ma.man.ac.uk/home/svn-oomphlib/public/trunk@232 595e1d82-adf1-4282-bac7-b6073fa2ca8e