Modified the build machinery so that we can build and link to Trilinos 10 as well as Trilinos 9. This is still not as nice as I would like because I haven't invested enough time into the autotools part. The script in external_distributions/trilinos will detect any file of the form trilinos*.tar.gz and extract the major version number to do the right thing (autotools vs cmake). I could not find an easy way of making autoconf use wildcards in the same way, so have hard-coded the .tar.gz files to be trilinos-9.0.2.tar.gz (as before) and trilinos-10.8.5-Source.tar.gz (the most current release as of 7th Feb 2012).
Needed one change in trilinos_solver.h. The latest version of Trilinos's
ML preconditioner uses the communicator of the matrix in its destructor, so
the preconditioner must be deleted before the matrix.
All tests pass on my machine with Trilinos 9 or Trilinos 10.

Also added LAPACK files clarnd.f and zlarnd.f that were 
required by Trilinos 10

At the same time, I've modified the double_vector_with_halo.h/.cc so that
the Halo/Haloed_equations are unsigned and passed between processors with
an MPI_UNSIGNED variable. Hopefully this is now robust.

I note that I haven't checked all possible configurations of 
Trilinos/BLAS/LAPACK, etc, so there may be an error, but I have tried to
mirror what was done with Trilinos 9, so I *hope* it'll all be OK. If not
feel free to shout!



git-svn-id: svn+ssh://vummath.ma.man.ac.uk/home/svn-oomphlib/public/trunk@264 595e1d82-adf1-4282-bac7-b6073fa2ca8e