Added the changes that *should* allow compilation without stacktrace.h (automagicaly detected)
Also added the FFLAGS_NO_OPT environment variable for use with blas library
when there are compilers that need flags other than -O0. This could probably
still be tidied up further. Changed all config files to include the new flag,
if it is not set then no arguments will be passed to the compiler when 
compiling the two files in the blas library.



git-svn-id: svn+ssh://vummath.ma.man.ac.uk/home/svn-oomphlib/public/trunk@277 595e1d82-adf1-4282-bac7-b6073fa2ca8e