matrices.cc/.h: Modified the concatenate_without_communication(...) function
1) Further generalised the concatenate_without_communication(...)
function so now it can concatenate matrices where the result is not square.
This relies on a Vector of distributions being passed for the block
columns, as well as a Vector of distributions for the block rows.
The Vector row_distribution_pt tells the function how to permute the rows.
The Vector col_distribution_pt tells the function how to permute the
columns.

2) Wrote a wrapper around the generalised version which takes only one
Vector of distribution to be used as both row_distribution_pt and
col_distribution_pt. This is so all current code will work, and it's
nicer to not have to pass the same argument twice if it is repeated.

3) Updated various PARANOID tests and added a few more!

4) Updated comments to reflect the change, and explain what
row_distribution_pt and col_distribution_pt is.

git-svn-id: svn+ssh://vummath.ma.man.ac.uk/home/svn-oomphlib/public/trunk@446 595e1d82-adf1-4282-bac7-b6073fa2ca8e