Minor tweaks to validation tolerances
Changed meaning of Problem::Bifurcation_detection
The sign change of the determinant of the Jacobian will only be monitored
during arc-length stepping if Problem::Bifurcation_detection = true 
(default is false).  
Added the additional flag Problem::Bisect_to_find_bifurcation which plays
the role previously assigned to Bifurcation_detection. It does as it says and
will adjust the arc-length to use bisection to find the point at which the
determinant of the Jacobian matrix is zero.
Note that Bifurcation_detection must be true in order for the 
Bisection method to work.

Fixed a mistake in the boundary coordinates of FaceElements in Telements.cc

Modified the spherical_navier_stokes elements so that the LSC preconditioner
still works with them!



git-svn-id: svn+ssh://vummath.ma.man.ac.uk/home/svn-oomphlib/public/trunk@474 595e1d82-adf1-4282-bac7-b6073fa2ca8e