Remove strange line from "config/configure.ac_scripts/end"
The line is "AS_DIRNAME([/* automatically generated */], $ac_prefix_conf_OUT)".

The autoconfig documentation says this should get the directory name of
        "/* automatically generated */"
which obviously is not a file...

Instead I think it attempts to get the dirname of everything in the glob
"/*", i.e. everything in your root dir. Depending on the version of
autotools this either gives an error or a (long) list of "/"s.

It can't possibly be doing anything useful because the result of the macro
isn't assigned to a variable. So I removed it.

git-svn-id: svn+ssh://vummath.ma.man.ac.uk/home/svn-oomphlib/public/trunk@479 595e1d82-adf1-4282-bac7-b6073fa2ca8e