Fix bug in variable_from_makefile in parallel_self_test.py
If there is no Makefile in the pwd then this function would always fail
because make fails (even though a Makefile in the pwd is not really
needed).

Fixed by setting the cwd for the make command to be that of the Makefile
from which the variable is being extracted.

git-svn-id: svn+ssh://vummath.ma.man.ac.uk/home/svn-oomphlib/public/trunk@487 595e1d82-adf1-4282-bac7-b6073fa2ca8e