Tweaked some of the validation parameters for Patrick's new demos
Adjusted the Makefile.am's so that the appropriate files get included in
make mydist *Only checked* Patrick's tests after distribution and untar-ing

Added some more modifications as part of the big continuation/bifurcation 
rewrite:

 Added a new boolean flag to set_timestepper() functions that specifies
whether existing data is to be wiped. The default behaviour is that only
the current value is preserved, but if the boolean is set to true then all
exising time history and auxillary data is preserved (or all that will fit
within the newly allocated storage)

Added new get_inner_product() functions to assembly handlers and elements that
are to be used in bifurcation location and continuation. The plan is to 
merge the required functionality with that in projection problems because the
information required is almost the same.



git-svn-id: svn+ssh://vummath.ma.man.ac.uk/home/svn-oomphlib/public/trunk@737 595e1d82-adf1-4282-bac7-b6073fa2ca8e