Allow make_node_periodic(..) to make copies of copies of nodes
In make_node_periodic(...) if the node to be copied is already a copy it
now handles it instead of crashing.

It handles it by making the node a copy of the "node_to_be_copied"'s
copied node (we need better terminology here... sorry).

So starting with n1 not a copy and n2 a copy of N:

n1
n2 -> N

if we call make_node_periodic(n1, n2) we get:

n1 -> N
n2 -> N

which seems sensible.

git-svn-id: svn+ssh://vummath.ma.man.ac.uk/home/svn-oomphlib/public/trunk@826 595e1d82-adf1-4282-bac7-b6073fa2ca8e