In FullCircleMesh add paranoid check of input sizes
Mesh constructor can only generate meshes of 5 elements, so the size of
the input vectors must be 4. This was not checked anywhere in the code,
and trying to create more refined meshes by adding new "theta_position"s
resulted in singular elements.

Added a paranoid check to stop other people wasting time on this.

git-svn-id: svn+ssh://vummath.ma.man.ac.uk/home/svn-oomphlib/public/trunk@879 595e1d82-adf1-4282-bac7-b6073fa2ca8e