In autogen.sh use autoreconf --force rather than deleting helper scripts
Deleting the scripts is problematic when we only want to generate the
config files because they may not be re-created, resulting in a dist tar
file which cannot be built without automake installed.

So instead we set the autoreconf --force flag to force regeneration of
the helper scripts. May also have additional benefits because all
autoconf files are regenerated, which is what we want when we specify -r
to autogen.sh.

git-svn-id: svn+ssh://vummath.ma.man.ac.uk/home/svn-oomphlib/public/trunk@1017 595e1d82-adf1-4282-bac7-b6073fa2ca8e