Next modification in free-surface capabilities. Introduction of  function that computes surface gradients and surface divergence in FluidInterfaceElement. This neatly packages away all the mathematics, so that the equations can be implemented generically.
Note that the Jacobian computed by this function includes the factor of r in Axisymmetric cases
which has corrected an oversight in the previous implementation (although it only changed
the definition of the fake Lagrange multiplier that enforced the kinematic condition
in ElasticMeshes so doesn't really matter, but it's nice to get it correct!).

The interfaces have changed sufficiently that anybody who actually used this will have to
work out what is going on because the code will fail to compile.

I should also have mentioned last time that the add_additional_values() function has
changed so that the number of values assigned to a particular ID can be resized but
only if the storage is at the end of the node's values.



git-svn-id: svn+ssh://vummath.ma.man.ac.uk/home/svn-oomphlib/public/trunk@1087 595e1d82-adf1-4282-bac7-b6073fa2ca8e