Finally oomph-lib works on macs again! The key is to specify the specification of the configure flag
   --disable-shared

because mac-osx struggles with shared libraries. 
There's a new configure options file, specifically for macs

   mac_osx_paranoid_debug_range_checking

(currently only for the "go extra slow" build options, but
it shows what needs to be done).

Also had to provide a dummy replacement for

    fpu_control.h

which triangle uses to get floating point characteristics
of the machine. Sadly this file doesn't exist on macs...

Also fixed various issues in David Shepherds rewrite of the
noninteractive autogen script that didn't work on macs.

Finally, fixed a problem with a header path in tetgen_scaffold_mesh.h
and the associated Makefile.am. Not sure why this used to work (or 
why it stopped now). 



git-svn-id: svn+ssh://vummath.ma.man.ac.uk/home/svn-oomphlib/public/trunk@1091 595e1d82-adf1-4282-bac7-b6073fa2ca8e