Fixed remaining clang warnings. Mostly uninitialised variables and hidden overloaded functions.
The most nasty was an issue to do with namespace encapsulation and friends.
Essentially, if a friend class is implicitly defined within a namespace then
the class definition must also be defined within the same namespace ... and
not the global scope. g++ incorrectly allows the definition to be at the 
global scope.



git-svn-id: svn+ssh://vummath.ma.man.ac.uk/home/svn-oomphlib/public/trunk@1163 595e1d82-adf1-4282-bac7-b6073fa2ca8e