Fixes for:
      - src/generic/problem.h
      - src/generic/problem.cc
      - demo_drivers/unsteady_heat/two_d_unsteady_heat_2adapt/validata/

In my big upload, I had somehow deleted the work that MH had added in revision
1337 (specifically, stuff related to Problem::Keep_temporal_error_below_tolerance).
I've added this code back in. This, in turn, affected the validata data file in two_d_unsteady_heat_2adapt, so that needed to be reset too (i.e. the one that I
had claimed was incorrect in revision 1351).

git-svn-id: svn+ssh://vummath.ma.man.ac.uk/home/svn-oomphlib/public/trunk@1352 595e1d82-adf1-4282-bac7-b6073fa2ca8e