Add further matrix functions.
Add two matrix add functions and a matrix multiply to the compressed row complex matrix class. The latter has 3 implementation options and the choice is hard coded. This code was implemented by Jack Keeler.

From supplementary material of the paper: 2019 ``The influence of invariant solutions in the transient behaviour of an air bubble in a Hele-Shaw channel", Keeler, J. S., Thompson, A. B. & Lemoult, G. & Juel, A. & Hazel, A. H., Proc. R. Soc. A, 20190434.