Get rid of some aborts in superlu_dist.c
Fixed overzealous catching of default cases in switch statements introduced in the last commit. It seems that nothing needs to be done for certain settings of the switch parameter, so the behaviour was correct; just dodgy/unclear code that the compiler rightly warned about. Left the aborts in in the cases that weren't visited so we can double check what's going on if the code ever gets there (have added some diagnostics to facilitate this).