Add integration element and two demos.
Gaussian demos integrate a Gaussian function over differently shaped
domains and dimensions.
Centre of mass demos calculate the centre of mass of different domains
in different dimensions.
Integral elements add the contribution from the integral.
Info elements provide storage for the result of the integration, or
another computation.
TIntegralElements also compute the integrand flux to provide flux for
the Z2ErrorEstimator to refine about localised changes in the integrand.