Change adaptive dt safety factor to error s.f.
Changed the safety factor that appears in the calculation of a new time
step to alter the error target relative to the error tolerance and
changed the name
    Adaptive_dt_safety_factor  -->  Target_error_safety_factor

Also, changed its default value to 1.0 so that it has no impact on the
current behaviour of oomph-lib unless the user specifies.

A more verbose doxygen comment has been given to the variable too which
justifies its use and suggests values commonly used in literature.