Remove pure virtual overrides from Z2 elements
This commit removes the pure virtual declarations of the functions

    virtual unsigned nvertex_node() const
    virtual void vertex_node_pt() const

from the ElementWithZ2ErrorEstimator class as both are already inherited
from the FiniteElement base class as broken functions such that if they
are ever used without being reimplemented they throw an OomphLibError.

The default FiniteElement behaviour is sufficient to ensure the functions
end up implemented before they are used, removing the pure virtual
override prevents ambiguity in the case of multiple/diamond inheritance.