Add a min_element_interior_angle variable to the TriangleMeshParameters.
This is used to set up the min_element_interior_angle before the
TriangleMesh is built.

Add a min_element_interior_angle_for_adaption variable to the
RefineableTriangleMesh.
This is used to set the minimum triangle angle within triangle
mesh generation for the subsequent adaptions.
This defaults to the one used for the initial mesh generation.

Pass the minimum angle into the RefineableTriangleMesh constructors
where the TriangleMeshParameters is not used.