Add a switch in the ANS face elements to use standard vs bulk element
indexing.
This means that we do not need to include additional eigen slip
elements.
The validata needs updating as the total velocity and total pressure is
now outputted rather than the finite element correction.
There is a small difference in the pressure, but the solutions look to
be valid when plotted. (ie. no singularity remaining and no pressure
spikes around augmented regions etc.)